Hack to fix Geotab API issue with decoding special characters.
RestClient encodes special characters in the URL it sends to Geotab.
For example, https://my.geotab.com/apiv1/Authenticate?userName=test%40gmail.com&password=123%2C%2F%5Eabc&database=fleetio

For some reason, Geotab's API can decode special characters in some
cases, but not in others. Since the password field is entirely user
created, and since having special characters in your password is a
best practice, this presents a problem. Geotab has been unable to
pinpoint why it sometimes works and sometimes fails, but enclosing
the entire password params in quotes is a hack that seems to fix it.
Unfortunately, this means that if someone ever uses quotes in their
password, it won't work, but having one or two special characters
that don't work is much better than having a whole bunch that might
not work.